Refactor out shared utils
